+++ mealkitary-api/build.gradle.kts
@@ -21,6 +21,7 @@ dependencies {
     implementation(project(":mealkitary-infrastructure:adapter-persistence-spring-data-jpa"))
     implementation(project(":mealkitary-infrastructure:adapter-paymentgateway-tosspayments"))
     implementation(project(":mealkitary-infrastructure:adapter-firebase-notification"))
+    implementation(project(":mealkitary-infrastructure:adapter-configuration"))
     implementation(
         project(
             ":mealkitary-infrastructure:business-registration-number-validator:adapter-open-api-brn-validator",
